edit-user - Updated edit user to match delete user controller style
