+++ src/main/routes.ts
@@ -2,6 +2,7 @@ import { Application } from 'express';
 import {
   ADD_USER_DETAILS_URL,
   ADD_USERS_URL,
+  EDIT_USER_URL,
   HOME_URL,
   MANAGER_USERS_URL,
   USER_ACTIONS_URL,
@@ -18,6 +19,7 @@ export default function(app: Application): void {
   app.get(ADD_USERS_URL, featureFlags.toggleRoute(BETA_FEATURES), app.locals.container.cradle.addUsersController.get);
   app.post(ADD_USER_DETAILS_URL, featureFlags.toggleRoute(BETA_FEATURES), app.locals.container.cradle.addUserDetailsController.post);
   app.get(MANAGER_USERS_URL, app.locals.container.cradle.manageUsersController.get);
+  app.post(EDIT_USER_URL, app.locals.container.cradle.userEditController.post);
   app.post(USER_RESULTS_URL, app.locals.container.cradle.userResultsController.post);
   app.post(USER_ACTIONS_URL, featureFlags.toggleRoute(BETA_FEATURES), app.locals.container.cradle.userActionsController.post);
 }
