+++ src/main/utils/urls.ts
@@ -9,6 +9,7 @@ export const ADD_USERS_URL = '/users/add';
 export const ADD_USER_DETAILS_URL = '/users/add/details';
 export const USER_DETAILS_URL = '/users/details';
 export const USER_ACTIONS_URL = '/users/actions';
+export const EDIT_USER_URL = '/users/edit';
 
 // Internal URLs
 export const USER_RESULTS_URL = '/user-results';
